requirements.txt
Flask
Flask-MySQLdb
Flask-WTF
pandas
scikit-learn
nltk
fuzzywuzzy
python-Levenshtein
mysql-connector-python
